Move geolocation stuff to bugs
